organizando estrutura para arquitetura hexagonal
