+++ src/main/java/com/github/lucasmartins/task_lister/application/core/usecase/LoginUseCase.java
@@ -1,0 +1,8 @@
+package com.github.lucasmartins.task_lister.application.core.usecase;
+
+import com.github.lucasmartins.task_lister.domain.model.User;
+
+public interface LoginUseCase {
+
+    User login(User user);
+}
