+++ src/main/java/com/github/lucasmartins/task_lister/domain/model/User.java
@@ -1,31 +1,20 @@
 package com.github.lucasmartins.task_lister.domain.model;
 
-import jakarta.persistence.*;
+import com.github.lucasmartins.task_lister.domain.model.abstractions.AbstractDomain;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.util.List;
-
+@Builder
 @Getter
 @Setter
-@Entity(name = "users")
-public class User {
+public class User implements AbstractDomain {
 
-    @Id
-    @GeneratedValue(generator = "user_id_sequence", strategy = GenerationType.SEQUENCE)
     private int id;
 
-    @Column(nullable = false)
     private String username;
 
-    @Column(nullable = false)
     private String password;
 
-    @Column(nullable = false)
     private String email;
-
-    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
-    private List<Goal> goals;
-
-
 }
