+++ src/main/java/com/github/lucasmartins/task_lister/domain/model/abstractions/AbstractDomain.java
@@ -1,0 +1,4 @@
+package com.github.lucasmartins.task_lister.domain.model.abstractions;
+
+public interface AbstractDomain  {
+}
