+++ src/main/java/com/github/lucasmartins/task_lister/domain/model/abstractions/AbstractEntity.java
@@ -1,0 +1,5 @@
+package com.github.lucasmartins.task_lister.domain.model.abstractions;
+
+public interface AbstractEntity {
+
+}
