--- src/main/java/com/github/lucasmartins/task_lister/domain/repository/GoalRepository.java
@@ -1,9 +1,0 @@
-package com.github.lucasmartins.task_lister.domain.repository;
-
-import com.github.lucasmartins.task_lister.domain.model.Goal;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface GoalRepository extends JpaRepository<Goal, Integer> {
-}
