--- src/main/java/com/github/lucasmartins/task_lister/domain/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.github.lucasmartins.task_lister.domain.repository;
-
-import com.github.lucasmartins.task_lister.domain.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-
-    Optional<UserDetails> findByUsername(String username);
-}
