+++ src/main/java/com/github/lucasmartins/task_lister/domain/service/LoginService.java
@@ -1,0 +1,20 @@
+package com.github.lucasmartins.task_lister.domain.service;
+
+
+import com.github.lucasmartins.task_lister.infrastructure.adapter.outbound.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class LoginService {
+
+    private final UserRepository userRepository;
+
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return userRepository.findByUsername(username)
+                .orElseThrow(() -> new UsernameNotFoundException("Usuário não encontrado"));
+    }
+}
