+++ src/main/java/com/github/lucasmartins/task_lister/infrastructure/adapter/outbound/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.github.lucasmartins.task_lister.infrastructure.adapter.outbound.repository;
+
+import com.github.lucasmartins.task_lister.infrastructure.persistence.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+
+    Optional<UserEntity> findByUsername(String username);
+}
