+++ src/main/java/com/github/lucasmartins/task_lister/infrastructure/config/security/service/CustomUserDetailsService.java
@@ -1,0 +1,23 @@
+package com.github.lucasmartins.task_lister.infrastructure.config.security.service;
+
+import com.github.lucasmartins.task_lister.infrastructure.adapter.outbound.repository.UserRepository;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    public CustomUserDetailsService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return userRepository.findByUsername(username)
+                .orElseThrow(() -> new UsernameNotFoundException("Usuário não encontrado"));
+    }
+}
