+++ src/main/java/com/github/lucasmartins/task_lister/infrastructure/persistence/GoalTypeEntity.java
@@ -1,0 +1,20 @@
+package com.github.lucasmartins.task_lister.infrastructure.persistence;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Entity(name = "goal_type")
+public class GoalTypeEntity {
+
+    @Id
+    @GeneratedValue(generator = "goal_type_id_sequence", strategy = GenerationType.SEQUENCE)
+    private int id;
+
+    private String description;
+}
