+++ src/main/java/com/github/lucasmartins/task_lister/infrastructure/persistence/TaskEntity.java
@@ -1,0 +1,40 @@
+package com.github.lucasmartins.task_lister.infrastructure.persistence;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDate;
+
+@Getter
+@Setter
+@Entity(name = "task")
+public class TaskEntity {
+
+    @Id
+    @GeneratedValue(generator = "task_id_sequence", strategy = GenerationType.AUTO)
+    private int id;
+
+    @Column(nullable = false)
+    private String title;
+
+    @Column(nullable = false)
+    private String description;
+
+    @ManyToOne
+    @JoinColumn(name = "task_category_id")
+    private TaskCategoryEntity taskCategoryEntity;
+
+    @Column(nullable = false)
+    private LocalDate dueDate;
+
+    @Column(nullable = false)
+    private LocalDate createdAt;
+
+    @Column(nullable = false)
+    private String updatedAt;
+
+    @ManyToOne
+    @JoinColumn(name = "user_id", nullable = false)
+    private GoalEntity goalEntity;
+}
